backend/requirements.txt
Django~=4.1.2
django-countries~=7.4.1
django-cors-headers~=3.13.0
djangorestframework~=3.14.0
drf-yasg~=1.21.4
environ~=1.0
psycopg2~=2.9.4
wheel~=0.37.1